Add CMainFrame::ResizePageInListbook; Add menu item: CMainFrame::OnAddPage, CMainFrame::OnDeletePage; Change menu item: CMainFrame::OnAddNewItem, CMainFrame::OnDeleteItem; Change CBasePanel::SaveXmlFileFromList; Change CPanel::GetValue: delete processing m_listctrl -> add processing m_grid; Add CPanel::GetSizeAllColumn, CPanel::SetSizeAllColumn; Change CPanel::AddNewItem and CPanel::DeleteItem for m_grid; Add helpers user interface

+++ MainFrame.cpp
@@ -4,11 +4,11 @@
 #define HEIGHT 500
 
 CMainFrame::CMainFrame(wxWindow* parent)
-    : wxFrame(parent, wxID_ANY, wxT("CodeManager"))
+    : wxFrame(parent, wxID_ANY, wxT("Code Manager"))
 {
     SetIcon(wxICON(lock));
 
-    m_choicebook = new wxListbook(this, NewControlId(), wxDefaultPosition, wxDefaultSize);
+    m_choicebook = new wxListbook(this, NewControlId());
 
     CreateMenuBar();
 
@@ -46,6 +46,8 @@ void CMainFrame::LoadXml()
             }
         }
     }
+
+    ResizePageInListbook();
 }
 
 void CMainFrame::OnSaveXml(wxCommandEvent& event)
@@ -69,6 +71,37 @@ void CMainFrame::OnSaveXml(wxCommandEvent& event)
     xmlDoc.Save(NAMEFILEXML);
 }
 
+void CMainFrame::ResizePageInListbook()
+{
+    wxArrayInt arraySizeCol;
+
+    for(unsigned int page = 0; page < m_choicebook->GetPageCount(); page++) {
+        wxWindow* windowPage = m_choicebook->GetPage(page);
+        if(windowPage != nullptr) {
+            CPanel* panelPage = dynamic_cast<CPanel*>(windowPage);
+            if(panelPage != nullptr) {
+                wxArrayInt new_arraySizeCol;
+                panelPage->GetSizeAllColumn(new_arraySizeCol);
+                if(new_arraySizeCol.size() == arraySizeCol.size()) {
+                    for(unsigned int i = 0; i < new_arraySizeCol.size(); i++)
+                        if(arraySizeCol[i] < new_arraySizeCol[i])
+                            arraySizeCol[i] = new_arraySizeCol[i];
+                } else
+                    arraySizeCol = new_arraySizeCol;
+            }
+        }
+    }
+
+    for(unsigned int page = 0; page < m_choicebook->GetPageCount(); page++) {
+        wxWindow* windowPage = m_choicebook->GetPage(page);
+        if(windowPage != nullptr) {
+            CPanel* panelPage = dynamic_cast<CPanel*>(windowPage);
+            if(panelPage != nullptr)
+                panelPage->SetSizeAllColumn(arraySizeCol);
+        }
+    }
+}
+
 void CMainFrame::CreateMenuBar()
 {
     wxMenuBar* menuBar = new wxMenuBar;
@@ -79,36 +112,77 @@ void CMainFrame::CreateMenuBar()
     wxMenuItem* save = new wxMenuItem(file, NewControlId(), wxT("Save"));
     file->Append(save);
 
+    wxMenuItem* addPage = new wxMenuItem(edit, NewControlId(), wxT("Add Page"));
+    wxMenuItem* deletePage = new wxMenuItem(edit, NewControlId(), wxT("Delete Page"));
     wxMenuItem* addNewItem = new wxMenuItem(edit, NewControlId(), wxT("Add new item"));
     wxMenuItem* deleteItem = new wxMenuItem(edit, NewControlId(), wxT("Delete item"));
+
+    edit->Append(addPage);
     edit->Append(addNewItem);
+    edit->AppendSeparator();
+    edit->Append(deletePage);
     edit->Append(deleteItem);
 
     menuBar->Append(file, wxT("File"));
     menuBar->Append(edit, wxT("Edit"));
 
     Bind(wxEVT_COMMAND_MENU_SELECTED, CMainFrame::OnSaveXml, this, save->GetId());
+    Bind(wxEVT_COMMAND_MENU_SELECTED, CMainFrame::OnAddPage, this, addPage->GetId());
+    Bind(wxEVT_COMMAND_MENU_SELECTED, CMainFrame::OnDeletePage, this, deletePage->GetId());
     Bind(wxEVT_COMMAND_MENU_SELECTED, CMainFrame::OnAddNewItem, this, addNewItem->GetId());
     Bind(wxEVT_COMMAND_MENU_SELECTED, CMainFrame::OnDeleteItem, this, deleteItem->GetId());
+
     SetMenuBar(menuBar);
 }
 
 void CMainFrame::OnAddNewItem(wxCommandEvent& event)
 {
-    wxWindow* windowPage = m_choicebook->GetPage(m_choicebook->GetSelection());
-    if(windowPage != nullptr) {
-        CPanel* panelPage = dynamic_cast<CPanel*>(windowPage);
-        if(panelPage != nullptr)
-            panelPage->AddNewItem();
+    if(m_choicebook->GetPageCount() != 0) {
+        wxWindow* windowPage = m_choicebook->GetPage(m_choicebook->GetSelection());
+        if(windowPage != nullptr) {
+            CPanel* panelPage = dynamic_cast<CPanel*>(windowPage);
+            if(panelPage != nullptr)
+                panelPage->AddNewItem();
+        }
     }
 }
 
 void CMainFrame::OnDeleteItem(wxCommandEvent& event)
 {
-    wxWindow* windowPage = m_choicebook->GetPage(m_choicebook->GetSelection());
-    if(windowPage != nullptr) {
-        CPanel* panelPage = dynamic_cast<CPanel*>(windowPage);
-        if(panelPage != nullptr)
-            panelPage->DeleteItem();
+    if(m_choicebook->GetPageCount() != 0) {
+        int selected = m_choicebook->GetSelection();
+        if(selected != -1) {
+            wxWindow* windowPage = m_choicebook->GetPage(selected);
+            if(windowPage != nullptr) {
+                CPanel* panelPage = dynamic_cast<CPanel*>(windowPage);
+                if(panelPage != nullptr)
+                    panelPage->DeleteItem();
+            }
+        } else
+            wxMessageBox(wxT("Select item"));
+    } else
+        wxMessageBox(wxT("No table for delete item"));
+}
+
+void CMainFrame::OnAddPage(wxCommandEvent& event)
+{
+    wxTextEntryDialog dlg(this, wxT("Enter name page"), wxT("Name new page"));
+    if(dlg.ShowModal() == wxID_OK) {
+        wxString name = dlg.GetValue();
+        m_choicebook->AddPage(new CPanel(m_choicebook, name), name);
     }
+}
+
+void CMainFrame::OnDeletePage(wxCommandEvent& event)
+{
+    if(m_choicebook->GetPageCount() != 0) {
+        int selected = m_choicebook->GetSelection();
+        if(selected != -1) {
+            wxMessageDialog dlg(this, wxT("Delete page"), wxT("Delete page?"), wxOK | wxCANCEL);
+            if(dlg.ShowModal() == wxID_OK)
+                m_choicebook->DeletePage(selected);
+        } else
+            wxMessageBox(wxT("Select page"));
+    } else
+        wxMessageBox(wxT("No page for delete"));
 }
